added ownership check
only author can edit and delete his blog

+++ controllers/blogController.js
@@ -10,15 +10,15 @@ exports.blog_get = (req, res, next) => {
   });
 };
 
-exports.blog_view_get = async(req, res) => {
+exports.blog_view_get = async (req, res) => {
   try {
-    const blog = await Blog.findById(req.params.id)
-    const comment = await Comment.find({blog: req.params.id})
-    res.json({blog, comment})
+    const blog = await Blog.findById(req.params.id);
+    const comment = await Comment.find({ blog: req.params.id });
+    res.json({ blog, comment });
   } catch (error) {
-    res.status(404).json({ error: error })
+    res.status(404).json({ error: error });
   }
-}
+};
 
 exports.new_post = async (req, res) => {
   try {
@@ -36,15 +36,21 @@ exports.new_post = async (req, res) => {
 
 exports.update_put = async (req, res) => {
   try {
-    const blog = {
-      author: req.user._id,
-      content: req.body.content,
-      title: req.body.title,
-      draft: req.body.draft || false,
-      _id: req.params.id,
-    };
-    await Blog.findByIdAndUpdate(req.params.id, blog, {});
-    res.json(blog);
+    const owner = await Blog.find({ _id: req.params.id, author: req.user.id });
+    if (owner.length > 0) {
+      // const blog = {
+      //   author: req.user._id,
+      //   content: req.body.content,
+      //   title: req.body.title,
+      //   draft: req.body.draft || false,
+      //   _id: req.params.id,
+      // };
+      // await Blog.findByIdAndUpdate(req.params.id, blog, {});
+      // res.json(blog);
+      res.json(owner);
+    } else {
+      res.json({ error: "You can't edit someone else's blog" });
+    }
   } catch (error) {
     res.status(404).json({ error: error });
   }
@@ -52,8 +58,13 @@ exports.update_put = async (req, res) => {
 
 exports.blog_delete = async (req, res) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id);
-    res.json({ message: "Post deleted successfully" });
+    const owner = await Blog.find({ _id: req.params.id, author: req.user._id });
+    if (owner.length > 0) {
+      await Blog.findByIdAndRemove(req.params.id);
+      res.json({ message: "Post deleted successfully" });
+    } else {
+      res.json({ error: "You can't delete someone else's blog" });
+    }
   } catch (error) {
     res.status(404).json({ error: error });
   }
@@ -65,9 +76,9 @@ exports.comment_new_post = async (req, res) => {
       author: req.user._id,
       blog: req.params.id,
       title: req.body.title,
-      content: req.body.content
-    }).save()
-    res.json(comment)
+      content: req.body.content,
+    }).save();
+    res.json(comment);
   } catch (error) {
     res.status(404).json({ error: error });
   }
